Fix bug causing 403 on signed requests
The format of RFC2822 is somewhat malleable, so including it in signed
headers can (and does) cause signature mismatches. We do include the
X-Amz-Date header, so requests are still properly limited to a date
range and can't be used again e.g. by reply attacks.

Change logic for building headers to add the Date header _after_ the
Authorization header has been generated. This prevents Date from being
part of the signed headers.

Fixes #8

Signed-off-by: Nick Stevens <nick@bitcurry.com>