crypto: use RustCrypto utils over deprecated rust-crypto
This change started due to my unfortunate exposure to a bug impacting
the ability for my project to target aarch64:
https://github.com/DaGenix/rust-crypto/issues/383.

In that issue (and in others) it is clear that rust-crypto is now
deprecated in favor of the RustCrypto crates (here we need sha2/hmac).
This change drops in this replacement for rust-crypto.

Signed-off-by: Paul Osborne <osbpau@gmail.com>