Presigned PUT requests should not encode path
Hi there, 

While trying to perform a presigned PUT upload, I noticed the following : 
- uploads to a path without a "/" work fine and uploads with a "/" in path fails. 

So I started digging around and it seems the path is url-encoded when performing a presign. Which, I think should not be the case... 
The presign request itself is working fine, the actual upload PUT request will fail with a 403. 

The suggested change actually corrects the problem for my storage object provider (locally on Minio and online at Scaleway).