0.27.0 (#218)
* bucket: add optional HTTP timeout for all bucket operations (#215)

And apply the new field in the attohttpc (blocking) and reqwest
backends.  Surf does not expose such an option, but it's also
not as useful in async code: the caller can await with timeout.

* Add a request timeout to credentials HTTP GETs (#214)

Attohttpc sets connection and read timeouts by default.  However, by
default, there is no timeout on writing the request, nor any
end-to-end timeout.  Let applications impose one if they want to.

* Implement CopyObject bucket operation (#211)

* Fix sync impl

* chore(deps): update dependencies. (#209)

* chore(deps): update dependencies.

* chore: bump aws-creds version to 0.26.1

* Bump all deps

* fix edge case of sorting query params for canonical query string in v4 signature (#205)

* Happy clippy

* Better ECS handling

* Implement presign_delete (#195)

* use h1-client-rustls for surf

* aws-creds 0.26.2

* 0.27.0

* Update README

Co-authored-by: Paul Khuong <pvk@pvk.ca>
Co-authored-by: Johan Thomsen <jot@internetgruppen.dk>
Co-authored-by: Oscar Rainford <oscar@fourbs.com.au>
Co-authored-by: Cory Snyder <cfs2008@gmail.com>
Co-authored-by: Florian Marending <florian@marending.dev>