If file is smaller than one chunk, don't initiate multi-part upload
There already was a special case in put_object_stream() for a small files,
where the multi-part upload was aborted if the input turned out to be
small enough to fit in one chunk. But we can do better than that, and not
initiate the multi-part upload in the first place, avoiding two round
trips to S3 (one to initiate, and another to abort the multi-part upload).

While we're at it, refactor the loop slightly.