cargo: Move `tokio-stream` create behind the `with-tokio` feature
`tokio-stream` has a dependency on `tokio` and is only used within the
`request` module, which depends on the `with-tokio` feature.  As such,
in order to completely prevent `tokio` from appearing in the crate
graph, this crate must be marked optional and enabled when `with-tokio`
is requested, omitted otherwise.

Also includes two drive-by typo-fixes in the features section of the
readme.