Minor compilation fixes
Firstly, observe that aws-creds doesn’t compile without default
features enabled.  This is because Credentials::refresh method which
calls Credentials::default which is only defined if http-credentials
feature is enabled.  Change the code (and related tests) so that the
former method is present only if that feature is enabled.

Note that even with that, rust-s3 doesn’t compile without default
features enabled.  This is because rust-s3 depends on aws-creds with
the above bug present.  This will be fixed once the dependency is
updated.

Secondly, fix minor compilation warnings caused by conditional
compilation leading to some dead code.