Refactor to aid upcoming change to HTTP request backend API.
The RequestImpl type alias for 1 of the 3 backends is now referenced
in only one place, a new Bucket::make_request method which everything
else that previously called RequestImpl::new now calls.

The constructors for each of these backends are sync, as constructors
should be. This will make it easier to change to a different backend
factory API. The only reason these were async anyway is so that
credentials_refresh could be called, but make_request can just do that.