Rearrange the order of some method in impl Bucket. No functional change.
The next change needs to split the Bucket type into different impl
blocks with different generics. This textually moves the some of the
methods so that they are grouped to fit together into the impl blocks
they will need to belong to. Nothing is added or removed.

This change exists only to make the diff of the next change easier
to follow.